Rewrite interfaces for harvest change

+++ interfaces/IUniV2Router.sol
@@ -18,12 +18,12 @@ interface IUniV2Router {
     ) external payable returns (uint256[] memory amounts);
 
     function swapExactTokensForTokens(
-        uint amountIn,
-        uint amountOutMin,
+        uint256 amountIn,
+        uint256 amountOutMin,
         address[] calldata path,
         address to,
-        uint deadline
-    ) external returns (uint[] memory amounts);
+        uint256 deadline
+    ) external returns (uint256[] memory amounts);
 
     function getAmountsOut(uint256 amountIn, address[] memory path)
         external
